+++ client/src/App.js
@@ -4,10 +4,10 @@ import { Switch, Route } from 'react-router-dom';
 import UserBar from './components/UserBar';
 import NavBar from './components/NavBar';
 import Footer from './components/Footer';
-import Landing from './pages/Landing';
-import Register from './pages/Register';
-import Login from './pages/Login';
-import NotFound from './pages/NotFound';
+import Landing from './containers/Landing';
+import Register from './containers/Register';
+import Login from './containers/Login';
+import NotFound from './containers/NotFound';
 
 function App() {
   return (
